pyproject.toml
[project]
name = "watsonx-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "ibm-watsonx-ai>=1.5.2",
    "modelcontextprotocol>=1.0.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.54.0",
]
